run app without errors, even if there's no BBG running
