+++ fdc3/src/main.tsx
@@ -1,5 +1,6 @@
 import ReactDOM from 'react-dom/client'
-import App from './App.tsx'
+import App from './App'
+import "./App.css";
 
 window.addEventListener("FSBLReady", () => {
   ReactDOM.createRoot(document.getElementById('root')!).render(
